// src/app/models/admin/capphephoatdongkhoangsan/cpkhaithackhoangsan/cpkhaithacloaikhoangsan.model.ts
export class OutputCpKhaiThacLoaiKhoangSanModel {
    idkhaithacloaikhoangsan: string;
    tenkhoangsan: string;
    idloaikhoangsan: string;
    tenloaikhoangsan: string;
    tennhomkhoangsan: string;
    idcapphepkhaithac: string;
    truluong: number;
    donvitruluong: string;
    siteid: string
}

export class InputCpKhaiThacLoaiKhoangSanModel {
    idkhaithacloaikhoangsan: string;
    tenkhoangsan: string;
    idloaikhoangsan: string;
    idcapphepkhaithac: string;
    truluong: number;
    donvitruluong: string;
}